add more custom Dracula highlights
I tried respecting the spec as much as possible:
https://spec.draculatheme.com/

The result is not perfect but still much better than what's available by
default.

As an aside, I'm a little frustrated by the current state of syntax
files in (neo)vim. There are much better alternatives to the default
files, but since they're not shipped with the editor, colorschemes can't
rely on their highlight groups being present. Here's hoping tree-sitter
will change that.